src/result/CVE-2021-29510/fastapi-backend-template/pyproject.toml
[tool.poetry]
name = "fastapi-backend-template"
version = "0.1.0"
description = ""
authors = ["wsa"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.103.2"
uvicorn = "0.20.0"
asyncpg = "0.27.0"
databases = "0.4.3"
python-dotenv = "0.18.0"
starlette-context = "^0.3.5"
PyJWT = "^2.6.0"
passlib = "1.7.4"
argon2-cffi = "21.3.0"
aio-pika = "^8.2.5"
aioboto3 = "12.0.0"
python-multipart = "0.0.5"
asynch = "0.2.2"
python-json-logger = "2.0.7"
PyYAML = "6.0.1"

[tool.poetry.group.dev.dependencies]
responses = "^0.23.3"
coverage = "^7.3.2"
freezegun = "^1.2.2"
isort = "^5.12.0"
pre-commit = "^3.5.0"
pycodestyle = "^2.11.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
